+++ changelog.md
@@ -13,10 +13,17 @@
 
 - Support Markdown format
 
-### 2024/11/30 - v.1.0.7
+### 2024/11/30 - v1.0.7
 
 - Change item menu shape
 - Add Share All menu
 - Add Re question menu
 - Add Web link function
 - Release Android Version
+
+### 2024/12/04 - v1.0.8
+
+- Change the chat view orientation
+- Add token usage 
+- Add Search function
+